pyproject.toml
[tool.poetry]
name = "smart-bloging-platform"
version = "0.0.1"
description = "Api for blogging app"
authors = ["takeRednotBlue <maxymklym1996@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "smart_bloging_platform"}]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.12.0"
python-dotenv = "^1.0.0"
fastapi = "^0.103.2"
uvicorn = {extras = ["standart"], version = "^0.23.2"}
email-validator = "^2.0.0.post2"
pydantic-extra-types = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.36.0"
pydantic-settings = "^2.0.3"
fastapi-jwt-auth = "^0.5.0"
sqlalchemy = "^2.0.22"
easy-open-ai = "^0.1.8"
psycopg = "^3.1.12"
greenlet = "^3.0.1"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
httpx = "^0.25.0"
pytest-asyncio = "^0.21.1"
aiosqlite = "^0.19.0"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = "."
norecursedirs = ["postgres_data"]
testpaths = ["tests"]
filterwarnings = [
    "ignore",
    "default:::src"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
